Move the logic to the chat model
